начали работу над выводом комментариев
